fix(demo): update codesandbox poc file structure for react-dom
